Add the feature of get end divce info
Add the feature of get end divce info by registering a macro of ZDO.Just
registering the Msg but not add the handle code.
This version is GW-withoutSTM32-beta0.13.This version is match up with
ECG-beta0.13.
The total feature is when the GW work key is pressed sending the "hello"
from GW to ECG.The ECG will show "hello" on OLED.